# api-contract.yaml
openapi: 3.1.0
info:
  title: CRM for Florists API
  description: API для системы управления цветочным бизнесом
  version: 2.0.0
servers:
- url: http://localhost:8011
  description: Development server
paths:
  /api/clients/:
    get:
      summary: Get Clients
      description: Получить список клиентов с фильтрацией
      tags:
      - clients
      operationId: get_clients_api_clients__get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: client_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Type
      requestBody: null
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
    post:
      summary: Create Client
      description: Создать нового клиента
      tags:
      - clients
      operationId: create_client_api_clients__post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
  /api/clients/{client_id}:
    get:
      summary: Get Client
      description: Получить клиента по ID
      tags:
      - clients
      operationId: get_client_api_clients__client_id__get
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      requestBody: null
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
    put:
      summary: Update Client
      description: Обновить клиента
      tags:
      - clients
      operationId: update_client_api_clients__client_id__put
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
    delete:
      summary: Delete Client
      description: Удалить клиента
      tags:
      - clients
      operationId: delete_client_api_clients__client_id__delete
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      requestBody: null
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
  /api/clients/{client_id}/orders:
    get:
      summary: Get Client Orders
      description: Получить заказы клиента
      tags:
      - clients
      operationId: get_client_orders_api_clients__client_id__orders_get
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: integer
          title: Client Id
      requestBody: null
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
...
  /api/inventory/:
    get:
      summary: Get Inventory
      description: Получить список складских позиций
      tags:
      - inventory
      operationId: get_inventory_api_inventory__get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: low_stock
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Low Stock
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      requestBody: null
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
    post:
      summary: Create Inventory Item
      description: Создать новую складскую позицию
      tags:
      - inventory
      operationId: create_inventory_item_api_inventory__post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
  /api/inventory/{inventory_id}:
    get:
      summary: Get Inventory Item
      description: Получить складскую позицию по ID
      tags:
      - inventory
      operationId: get_inventory_item_api_inventory__inventory_id__get
      parameters:
      - name: inventory_id
        in: path
        required: true
        schema:
          type: integer
          title: Inventory Id
      requestBody: null
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
    put:
      summary: Update Inventory Item
      description: Обновить складскую позицию
      tags:
      - inventory
      operationId: update_inventory_item_api_inventory__inventory_id__put
      parameters:
      - name: inventory_id
        in: path
        required: true
        schema:
          type: integer
          title: Inventory Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Inventory'
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
    delete:
      summary: Delete Inventory Item
      description: Удалить складскую позицию
      tags:
      - inventory
      operationId: delete_inventory_item_api_inventory__inventory_id__delete
      parameters:
      - name: inventory_id
        in: path
        required: true
        schema:
          type: integer
          title: Inventory Id
      requestBody: null
      responses:
        '200':
          description: Successful Response
        '422':
          description: Validation Error
...
components:
  schemas:
    Client:
      properties:
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        phone:
          type: string
          title: Phone
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        client_type:
          allOf:
          - $ref: '#/components/schemas/ClientType'
          default: оба
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - phone
      title: Client
      description: Модель клиента (заказчик/получатель)
    ClientType:
      type: string
      enum:
      - заказчик
      - получатель
      - оба
      title: ClientType
...
    Inventory:
      properties:
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        name:
          type: string
          title: Name
        quantity:
          type: number
          title: Quantity
        unit:
          type: string
          title: Unit
        min_quantity:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Quantity
        price_per_unit:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Per Unit
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - name
      - quantity
      - unit
      title: Inventory
      description: Модель складского учета
...